Faster computation of net worth and monthly balances
This eliminates use of summarize.clamp_opt, which must process all
transactions prior to the begin date in order to synthesize opening
balance transactions.  This led to (approximately) quadratic time
complexity in the length of the journal to compute monthly information.

Instead, for monthly balances, where the cumulative balance is not
required, we just limit the list of transactions by efficient binary
search, and for net_worth, where it is required, we compute it
incrementally.