Make JavaScript-includes DRY and remove packing of JavaScript-files (close #50)
The packing of JavaScript-files into one does pose a permissions-üroblem
when beancount-web is installed as one user, and run as another. At
runtime webassets creates the concatenated file, but as it is run under
a different user that it was installed by it has no permission to write
those generated files to disk, generating a 500. This changeset removes
this feature altogether, serving the JavaScript-files as single files.
This should not have a big performance-drawback, as it is run locally
anyway, and and the files get cached by the browser.