Don't use background-task watcher (#56)
The background-task watcher for file changes in livereload appears to be broken (and not imported correctly in the previous change)

Some text editors will save files as "move-original, create-new, delete-original" instead of "modify-original". This causes Inotify to loose track of the source file and stop reloading. Re-watch the sources files after every change, so Inotify doesn't loose them.