set page title in _layout; reduce template context
This keeps (almost) all the page names in one file and allows to check for
consistency more easily. Still a bit against DRY, since the pages are
listed twice in the layout (once for the nav-bar, once for the title),
but I guess better than before.

API attributes that can be easily accessed don't need to passed to the
template separately.