Downloads: Correct sheet order
Problem:
In the downloads for XLS, XLSX and ODS the "Query" is the first
sheet, "Results" the second. This is confusing.

Root cause:
When constructing the `pyexcel.Book` a `dict` is provided with
the required sheets. A dict is unordered, as as result, the
Sheets are added in alphabetical order to the Book.

Solution:
Use an OrderedDict, so the correct ordering is maintained.