Shutdown LiveReload before saving source files (#197)
Problem:
When LiveReload is active, saving a source file caused LiveReload
to reload the page. As a result, the cursor position got lost.

Solution:
Stop LiveReload before saving. To do this, one needs to dispatch
as `LiveReloadShutDown` event on the document. For some reason I
couldn't get this to work using jQuery's `trigger` or `triggerHandler`
functions, so I lifted the code used by LiveReload itself.