Changes to support low-res modes again (#358)
* Limit the height of the upper title text

So that it doesn't cover the menu button at very small page widths.  Uses
flexboxes to remove the need for padding, making the text fit nicer no matter
what.

* Fix chart hidder position when window width < 560px

The container wasn't positioned relatively, so the absolute position within it
was relative to the browser edge, not the article edge.  This lead to funny
business when smaller than 560px.