Don't toggle metadata indicator in Journal view
Because the class of the indicator was the same as the metadata-rows,
it's visibility was toggled with the metadata-rows in the Journal view.
This change adapts the class, so the indicators are always visible, even
when the metadata-rows are not shown/hidden by toggling them with the
keyboard shortcut or filter-buttons.