Don't load statement documents via the router mechanism
Statement documents like PDFs were, if clicked on the statement link in
the Journal view, loaded into the current page, which lead to weird
characters (the bytes of the document itself) being displayed. This
change makes the router mechanism ignore links with the class
"external", which is applied to the statement links.