Aggregate commodities when building an Inventory for net worth
Instead of tracking each lot individually and then reducing it to the target
currency, pretend that every posting belongs to the same lot. This reduces the
number of Positions in the Inventory significantly, saving time during the
reduce step. Since reduce happens for every date step in every operating
currency, this leads to significant reductions in net worth calculation.