Propagate end date to balance tree
Tree and TreeNodes are used to render the balance sheet and they have no access to the global end date filtering. We need a way to specify the date to look up prices for calculating market value. Use the transaction date of the last entry in the TreeNode to calculate the price.