Fix generated metadata uploading statements
The `statement` metadata key is not recognized as special by fava
anymore, since it has been renamed to `document`.  However, fava was
still inserting the now-obsolete `statement` metadata when uploading a
document by dropping it on a transaction in the journal view.

This patch makes fava add a `document` metadata instead of a `statement`
metadata.

Fixes #860