Fix Serialization of Balance.amount.number field (#967)
The current serialization of a Amount field in a entry like "...... balance .... 204.2 CHF" results in a
in a json object {"amount": 202.4, "currency": CHF"). While this is technically correct most
json parsers treat json numbers as floating points which can cause problems later on (see #964).

We replace amount with a dictionary to force serialization of the amount field as a string instead
(i.e json object will be .{"amount": "202.4", "currency": CHF"}.

This test would have failed without the serialisation patch.
We use a large decimal that cannot be represented in double
precision floating point to ensure that no floating point
conversions are being used during the serialisation/deserialisation
process.