Escape filter when clicking on payee in the journal
As filters use regular expression matching, we need to escape certain
characters in the payee string when adding a filter for it after
clicking on it in the journal.

This commit adds an "escape" parameter to addFilter() in case any future
filters need escaping also. As "escape" defaults to false, there is no
impact to old code.