Minor metadata filter improvements (#1114)
* Remove leading space in filter

* Split out journal filter escape function

It is sometimes necessary to escape only a part of the filter, so this
functionality needs to be exposed to the caller.

* Escape and anchor filter patterns

Previously, the payee name would be escaped, but not anchored, so a
shorter payee could also match a longer one. Metadata values were
neither escaped nor anchored. With this change, both types of filters
show only exact matches.

* Allow filtering by metadata key

* Add test for filtering posting metadata

* Filter for posting metadata using any

* Show indicators for posting metadata