allow scrolling of sidebar
The `overflow-y: auto` was removed at some point to allow the account
autocomplete list to overflow the aside. Since we cannott have the aside
scroll vertically but overflow horizontally, we have to put the
autocomplete list in fixed position. This means it will be incorrecly
positioned when the sidebar is scrolled, but that is preferable to the
sidebar not scrolling at all.

Fix #1387