explicit filtered ledger
The Ledger class, which contains most information about a given
Beancount file was previously always modified in-place to filter for the
currently given filters of the Fava frontend. With multiple requests
coming in, this lead to race conditions which resulted in incorrect data
being returned.

The Ledger class is not modified anymore to filter, but rather a new
class FilteredLedger is created for each requests. By using caching,
this should be more performant than the previous solution.

The attributes on the Ledger class that changed on filtering are now all
deprecated and do no respect the filters anymore. Extensions using them
can be fixed by accessing the attributes of `g.filtered` instead, e.g.,
`g.filtered.entries` instead of `ledger.entries`.