charts: refactors; move more logic to bar.ts
Generate stacks in bar.ts to allow for this logic to be tested. With
this we can also avoid some unnecessary calculations. For example, the
extent calculation is now trivial, since we do not separately need to
sum up the values but can rather use the already computed stacks.

In some places, the bar chart tried to be somewhat generic - it has
become very Fava-specific though (budgets, account stacks), so this
commit also changes some of the notation to be easier to understand
('currency' instead of 'name' for instance).