Set article overflow to scroll if larger than view port
On mobile the pages are often too wide for the viewport. For example
on frontpage the tree-table row usually don't fit on mobile to screen. This breaks
the whole layout by making the whole app to be scrollable sideways. This
especially breaks dialogs on mobile.

Solve this by setting maximum width for
articlte to the viewport width and make *them* scrollable instead of the
whole body.

Possible better alternative would be to add this style for
rows and few other specific wide elements so that only the wide elements
themselves would be side scrollable. However, this is "single place fix"
which fixes the immediate problem.