Prevent body scroll when dialog is open
Especially on mobile it might be confusing for user if the body scroll
is active when there is fullscreen dialog open. Either it will mess up
scrolling of the dialog content or when user closes the modal, their
location on the main page has changed seemingly randomly.

Solve this issue by manipulating body element's overflow style
programmatically when dialog is opened and closed.