Use 16px font on mobile to prevent form zooming
At least iOS Safari will zoom the view to every input if their font size
is less than 16px. This causes problem for example in the entry dialog
modal. When dialog is opened, the browser immediately zooms in placing
the closing button outside of the view. Therefore user has to scroll
horizontally or zoom out to again see the whole layout. Fix this by
using 16px fonts on mobile.

Note that this increases all of the fonts. While this might be good for
the readability of small displays, it also means less space for
everything. We might want to consider increasing only input fonts on
mobile.