for pages with exactly two graphs: fix graph cycling shortcuts (#1683)
fixes #1677

When a page has exactly two graphs, the two shortcuts for cycling to the
previous and next graph need to be bound to the same graph. This is one
way to allow this.

However, only the last instance of use:keyboardShortcut will show up in
the HTML attribute data-key. That’s why I have chosen this
shortcutForward() to be the last of the two.

Co-authored-by: Jakob Schnitzer <mail@jakobschnitzer.de>