core.file: improve newline handling and tests
Improve some of the fava.core.file tests and only write to ledgers in
tmp_paths explicitly created for tests to avoid flakiness related to
that.

Also, use the newline character already present in files when writing to
those files.

Also refactor the Watcher API to prepare for the watchfiles based file
watcher. Since these file change events are asynchronous, a method to
notify the file watcher of file changes that we trigger (all in
fava.core.file) is also added.