Allow entry to be empty for automatic post
This allows the account entry of a posting to be empty. This is
necessary for the automatic posting creation without passing a
boolean to AccountInput determining if it is the last posting to
save the transaction.

Without this check, the last posting will always be empty, hence the
transaction will not be saved.