automatically add posting field when last posting is non-empty (#1762)
* Remove + and p buttons on Transaction screen

Since new postings are automatically added, remove the + and the p
button on the transaction screen.

* Automatically add posting field when last posting is non-empty

This commit checks if information has been added to the last posting
when entering a transaction and adds an empty posting in this case.

* Removes two postings default

Since postings are automatically added on entering text, the default
number of postings need not be 2. This commit reduces the default number
of postings to 1.

* Allow entry to be empty for automatic post

This allows the account entry of a posting to be empty. This is
necessary for the automatic posting creation without passing a
boolean to AccountInput determining if it is the last posting to
save the transaction.

Without this check, the last posting will always be empty, hence the
transaction will not be saved.

* entry-form: minor fixes; ignore empty postings on serialisation
* entry-form: add JS class for Posting

Co-authored-by: Jakob Schnitzer <mail@jakobschnitzer.de>