fix: Respect ordering of provided ledgers
Previously the _add_env_filenames function converted the filenames to a
set. As [sets are unordered](https://docs.python.org/3/tutorial/datastructures.html#sets),
this prevented specifying order. Some bits of the code, notably the
default "/" route, relied on order to determine which the default ledger
was

The use of set() was first introduced in f33a63e35a1751001ac10fb992aba4e434ea7321.
It's not entirely clear from the commit, but presumably it was intended
to filter out duplicate items as this is a common pattern in Python. The
type is hinted to be an Iterable[] and lists are used in other places,
such as tests, so nothing requires using a set

To make test changes as minimal as possible, this commit makes the files
included from environment variables be ordered *first*

Close #1833