query: remove unused endpoint, add svelte-component for table
The query_result returns the rendered HTML which is now not used
anymore. Also remove the query charts methods that are now not needed
anymore.

Also remove the querytable.download_links macro - according to a GH code
search, this is not used anywhere else.

Add a svelte-component and a method on query_shell to more easily get
the data for it and use those two parts in the statistics report.