add support for Beancount v3
This uses beanquery to replace the functionality in beancount.query,
which works with both Beancount v2 and v3 - however there are some
differences on columns in beanquery compared to beancount.query.

For importers, this sticks to beancount.ingest on v2 and used beangulp
for v3. It still expectes importers to conform to the v2 importer
protocol for now. Due to changes related to duplicate detection,
duplicate detection is not automatically done by Fava here but can still
be manually specified as hooks.