feat: add fortnightly interval
In some countries (such as Australia), it is very common for wages to be
paid on a fortnightly basis, that is, every 2 weeks. As a result, a
number of other expenses are also paid on a fortnightly basis.

This commit extends the `Interval` enum by adding a `FORTNIGHT` option.
I have added a test for that, and have updated the translations where
possible (not that other than for English and French, the other
translations may be suboptimal).

Resolves: #1939
Signed-off-by: JP-Ellis <josh@jpellis.me>