fix: iso week usage
Fava appears to use ISO 8601 weeks (e.g. `2025-W01`), but unfortunately
does so incorrectly due to a mismatch between ISO 8601 and Gregorian
calendar week definitions.

ISO 8601 defines precisely how to handle weeks, namely, that the first
week of a year is the week that contains 4 January. This definition
results in some occasional mismatches with the Gregorian calendar such
as:

- The first week of 2025 starts on 30 December 2024
- 3 January 2021 belongs to Week 53 of 2020

Fortunately, the fix is straightforward as `strftime` provide[^posix]:

- `%G` for the week-based year
- `%V` for the week number of the year (with no 'week 0' as `%W` can
  return).

This change will ensure that there is no ambiguity on weeks that
straddle the new year, though at the cost of introducing occasional
off-by-one errors compared to previously generated data.

[^posix]: See [POSIX.1-2024](https://pubs.opengroup.org/onlinepubs/9799919799/functions/strftime.html)
Signed-off-by: JP-Ellis <josh@jpellis.me>