entries: immutable and migrate to runes mode
Svelte 5 in runes modes removes the possibility to have these classes be
automatically determined as mutated. It would have been possible to use
`$state` runes on the fields but using the runes in "normal" TS files
makes interop with plain Javascript harder to do and understand. So
getters/setters on the binds it is instead.

Also refactor the import report to separate component state and
(backend-provided) component properties.