styles: use light-dark() CSS function; convert to hsl(); simplify
This allows declaring colors for light and dark mode in the same place
and will also allow overriding the color-scheme centrally without having
to duplicate styles. This also made it easier to identify some color
vars that could be simplified / cleaned up.

Also, convert all colors to hsl() colors which makes it easier to get
matching (but darker colors) by only adjusting the luminance.

Try to adjust contrast for better visibility in places as well.