fix: correctly navigate to chosen line in editor
The $searchParams store value was only updated after navigation finished
so the initial run on opening a file would not successfully get the line
from the URL query string. So add the line number to the data extracted
in the route load function and use that value to scroll to the given
line.