lint: add biome config and use if for some filetypes
Biome does not support svelte template or yaml formatting so keep
prettier in place for those. For Typescript there was only one place
where the formatting changed, so it is very close to prettier.

Performance is great and it brings some useful linting rules as well.
For linting we still keep eslint though, it is no match for the
type-aware rules of typescript-eslint.