Hierarchical file menu for editor
Replace the editor File menu with a hierarchical menu, which resembles
a file picker in a regular text editor.

There is a subtle change from previous behaviour which is that the
currently selected file will be highlighted by having a different
background color for the currently selected file in the menu. This is
instead of the caret '>' in front of the file name in the current
version.

Otherwise, for users with only a few beancount files, the menu will look
almost like it used to. The only difference will be that the directory
where the file(s) reside is shown in the top and the files then listed
below.

So instead of:
```
/beancount/data/root/main.bean
/beancount/data/root/other.bean
```

It will look like:
```
/beancount/data/root
 main.bean
 other.bean
```

For users with multiple files in a directory hierarchy, the change will
be bigger.

Instead of:
```
/beancount/data/root/main.bean
/beancount/data/root/other/deeper/filec.bean
/beancount/data/root/other/deeper/filed.bean
/beancount/data/root/other/filea.bean
/beancount/data/root/other/fileb.bean
```

It will look like:
```
/beancount/data/root
 main.bean
 ▸ other/
```

And can be expanded to:
```
/beancount/data/root
 main.bean
 ▾ other/
   ▾ deeper/
     filec.bean
     filed.bean
   filea.bean
   fileb.bean
```

Unlike usually directories are _not_ first in the order. Instead the
order of the files and directories follows the existing alphabetic file
ordering.