charts: deduplicate and only send hierarchy data once
Since the account trees and the hierarchy charts share the same data
types, we can avoid sending both and instead add these charts in the
frontend code.

This also ensures that the tree tables and the charts are consistent as
they are based on exactly the same data. For example on the account
report, the hierarchy charts on the "Balances" subreport were not using
the cumulative balances, this is now fixed.