json: use more compact representation
This considerably reduces the size of the JSON responses and should also
be more performant. Browser devtools usually handle JSON natively and
can pretty-print it, so this should not make it harder to debug these
endpoint responses.

To still have more readable snapshots, these JSON files are not compared
as strings anymore but on the loaded contents, making the JSON snapshot
files indentation-insensitive. These snapshots are now pretty-printed
with biome just like any JSON file in the repository would be.