journal: render report in frontend, fetch rendered journal via API
This simplifies a lot of the involved logic since it more fully happens
in the well-type-checked Typescript code and Svelte components, so
there is less manual work to read attributes and add event listeners.

Also makes it easy to fetch them ordered by date either in ascending or
descending. Disallow sorting while the page is still loading to not
overwhelm the browser unnecessarily.