document upload: improve form, reduce HTML attribute footprint
In the journal, we had attributes for every posting for entry hashes and
dates - this can be dynamically inferred when a drop happens. This now
supports all entry types (that have accounts, not just transactions).

Reformat and re-align the document upload modal, add a file `<input>`
that allows changing the uploaded files.

Since dropping on e.g. a generated document entry would lead to garbage
being inserted, check for lineno=0 and raise an error in that case.