ci: apply some zizmor suggestions; remove caching
Disable caching in publish workflows for security. Also disable uv and
node caching in all builds - while it's most critical in the publish
workflow, for consistency (and because it does not seem to make a big
difference in runtime), disable it for all builds

Use gha-update instead of self-written script to update used github
actions and pin the versions to sha hashes.