Extend import help, add currency/commodity synonym, link to beancount option help
This fixes a few issues that were hampering me as a newcomer to
understand the import process.

The import help text is less dense now and presupposes less knowledge
about importing in beancount and the detailed differences of the
different generations of import mechanisms (beancount.ingest vs
beangulp). There are clearer step-by-step instructions and an example
that should get a user floating for a basic import-CSV case. Also, the
full signature of hook function is now documented, which wasn't done
anywhere in the documentations of beancount, beangulp, or fava.

Additionally, this adds to the syntax overview the simple information
that commoditys = currencies from the perspective of beancount. This
would have helped me as a newcomer to find correct BQL functions for my
needs.