fix: update narration tags/links on Enter key press (#2165)
* Revert "fix: narration isn't stored when the focus is on the narration input (#2149)"

This reverts commit fd252134a2be0718ecf66f43abd4f93e9b113c45.

This commit caused a regression where spaces could not be entered in the
narration input field. Reverting to fix the space input issue and will
implement a proper fix.

* fix: update narration tags/links on Enter key press

Add onEnter handler to narration input in Transaction.svelte to ensure
narration tags and links are processed when the user presses Enter.